iqtune app: wait for stats to be available before reading them
In order to read all the stats (up/down x bin/avg), we change the
statistic-profile to "full" for a short duration.
After having done that, we need to wait some frames to give time to
the kernel to process a full stat gathering cycle.
If we do not wait, we read obsolete stats.

Signed-off-by: Fabien Dessenne <fabien.dessenne@st.com>